// src/types/proveedores.types.ts
export interface ProveedorDto {
  prvPrimarykeyMprv?: number;
  prvIdentifkeyMprv: string;
  prvNumeronitMprv: string;
  prvRazonsocialMprv: string;
  prvObjetosocialMprv?: string;
  sisTiposociedadTpso?: string;
  sisCodactividadCiiu?: string;
  prvFechconstMprv?: string;
  prvPaginawebMprv?: string;
  prvDireccionMprv?: string;
  prvTelefonoMprv?: string;
  prvCorreoMprv?: string;
  sisCodpaiSipa?: string;
  sisIdedptSidp?: string;
  sisCodproSipr?: string;
  prvCodposMprv?: string;
  prvIdentifkeyRelg?: string | null;
  prvEstadoregMprv?: string;
}